+++ scripts/generate-website-yaml.py
@@ -20,8 +20,8 @@ from a local computer to access the database
 sqlalchemy engine. It should not be committed to GitHub.
 """
 
-# out_path = "C:/Users/sam.woodman/Downloads"
-out_path = "../glider-lab-manual/content"
+# yaml_path = "C:/Users/sam.woodman/Downloads"
+yaml_path = "../glider-lab-manual/content"
 
 if __name__ == "__main__":
     logging.basicConfig(
@@ -31,36 +31,31 @@ if __name__ == "__main__":
     )
 
     with open("db/glider-db-prod.txt", "r") as f:
-        conn_string = f.read()
-    
-    engine = sqlalchemy.create_engine(conn_string)
-    with engine.connect() as connection:
-        # Generate deployment table
-        df_depl = config.make_deployment_table(connection)
-        df_depl = df_depl.drop(["Dates", "Sensors"], axis=1)
+        conn_string = f.read()    
+        engine = sqlalchemy.create_engine(conn_string)
+        with engine.connect() as connection:
+            # Generate deployment table
+            df_depl = config.get_deployment_table(connection)
+            x = df_depl.copy(deep=True)
+            x = x.drop(["Dates", "Sensors"], axis=1)
 
-        # df_depl.to_csv(
-        #     "C:/Users/sam.woodman/Downloads/fleet-status-deployments.csv",
-        #     index=False
-        # )
+            # Write Deployments table to fleet status
+            wk_name = "Deployments-Database"
+            logging.info("Updating the Fleet Status %s sheet", wk_name)
+            x = x.fillna("").rename({"Glider_Deployment_ID": "Deployment_ID"})
+            gc = gspread.oauth()  # type: ignore
+            sh = gc.open("Fleet Status")
+            wk = sh.worksheet(wk_name)
+            wk.update([x.columns.values.tolist()] + x.values.tolist())
 
-        # Write Deployments table to fleet status
-        wk_name = "Deployments-Database"
-        logging.info("Updating the Fleet Status %s sheet", wk_name)
-        df_depl = df_depl.fillna("").rename({"Glider_Deployment_ID": "Deployment_ID"})
-        gc = gspread.oauth()  # type: ignore
-        sh = gc.open("Fleet Status")
-        wk = sh.worksheet(wk_name)
-        wk.update([df_depl.columns.values.tolist()] + df_depl.values.tolist())
+            # # Update data validation formatting automatically..
+            # from gspread.utils import ValidationConditionType
+            # wk.add_validation(
+            #     f'F2:L{1+x.shape[0]}',
+            #     ValidationConditionType.one_of_list,
+            #     ['TRUE', 'FALSE'],
+            #     showCustomUi=True
+            # )
 
-        # # Update data validation formatting automatically..
-        # from gspread.utils import ValidationConditionType
-        # wk.add_validation(
-        #     f'F2:L{1+df_depl.shape[0]}',
-        #     ValidationConditionType.one_of_list,
-        #     ['TRUE', 'FALSE'],
-        #     showCustomUi=True
-        # )
-
-        # Website yaml
-        config.make_website_yaml(connection, out_path)
+            # Make website yaml
+            config.make_website_yaml(df_depl, yaml_path)