+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- (in progress) Updating config website yaml
-- Changed config module functions that scrape info from the database to all expect a `sqlalchemy.engine.Connectable` object as the first input
+- Changed config module:
+    - Functions that scrape info from the database to all expect a `sqlalchemy.engine.Connectable` object as the first input
+    - Changed `instrument_attrs` to `_get_instrument_attrs`,  and `make_deployment_table` to `get_deployment_table`, for consistency with functions from other modules
+    - `make_website_yaml` now takes a dataframe (the output of `get_deployment_table`) as an input, rather than a database connection
+    - The `make...yaml` functions now return the dictionary rather than the file path
+    - `_read_esdglider_yaml` is now a full internal function
 - Changed plots to use a set legend position when possible, and to remove the 'inflections' tvt plot as it is functionally a duplicate of the 'diveEnergy' tvt plot
 - Updated `glider.binary_to_nc` to make the gridded data files using `glider.make_gridfiles_depth_measured` if the argument 'sci_timeseries_pyglider' is True.